Update git_walker specs to new format
