+++ utils/refactor_diligence.rb
